+++ SpatialGDK/Source/SpatialGDK/Private/LoadBalancing/GridBasedLBStrategy.cpp
@@ -5,6 +5,7 @@
 #include "EngineClasses/SpatialNetDriver.h"
 #include "EngineClasses/SpatialWorldSettings.h"
 #include "Utils/SpatialActorUtils.h"
+#include "Utils/SpatialStatics.h"
 
 #include "Templates/Tuple.h"
 
@@ -204,11 +205,8 @@ void UGridBasedLBStrategy::PostEditChangeProperty(FPropertyChangedEvent& Propert
 			const UWorld* World = GEditor->GetEditorWorldContext().World();
 			check(World != nullptr);
 
-			const ASpatialWorldSettings* WorldSettings = Cast<ASpatialWorldSettings>(World->GetWorldSettings());
-			check(WorldSettings != nullptr);
-
 			const UAbstractSpatialMultiWorkerSettings* MultiWorkerSettings =
-				WorldSettings->MultiWorkerSettingsClass->GetDefaultObject<UAbstractSpatialMultiWorkerSettings>();
+				USpatialStatics::GetSpatialMultiWorkerClass(World)->GetDefaultObject<UAbstractSpatialMultiWorkerSettings>();
 
 			for (const FLayerInfo WorkerLayer : MultiWorkerSettings->WorkerLayers)
 			{