+++ SpatialGDK/Source/SpatialGDKFunctionalTests/Private/SpatialFunctionalTestFlowController.cpp
@@ -9,7 +9,6 @@
 #include "LoadBalancing/LayeredLBStrategy.h"
 #include "Net/UnrealNetwork.h"
 #include "SpatialFunctionalTest.h"
-#include "SpatialFunctionalTestGridLBStrategy.h"
 #include "SpatialGDKFunctionalTestsPrivate.h"
 
 ASpatialFunctionalTestFlowController::ASpatialFunctionalTestFlowController(const FObjectInitializer& ObjectInitializer)
@@ -42,8 +41,12 @@ void ASpatialFunctionalTestFlowController::GetLifetimeReplicatedProps(TArray<FLi
 
 void ASpatialFunctionalTestFlowController::OnAuthorityGained()
 {
-	bReadyToRegisterWithTest = true;
-	OnReadyToRegisterWithTest();
+	// Super hack
+	FTimerHandle Handle;
+	GetWorldTimerManager().SetTimer(Handle, [this](){
+		bReadyToRegisterWithTest = true;
+		OnReadyToRegisterWithTest();
+	}, 0.5f, false);
 }
 
 void ASpatialFunctionalTestFlowController::Tick(float DeltaSeconds)
@@ -57,50 +60,6 @@ void ASpatialFunctionalTestFlowController::Tick(float DeltaSeconds)
 void ASpatialFunctionalTestFlowController::CrossServerSetWorkerId_Implementation(int NewWorkerId)
 {
 	WorkerDefinition.Id = NewWorkerId;
-}
-
-void ASpatialFunctionalTestFlowController::AddEntityInterest_Implementation(const int64 ActorEntityId)
-{
-	ChangeEntityInterest(ActorEntityId, true);
-}
-
-void ASpatialFunctionalTestFlowController::RemoveEntityInterest_Implementation(const int64 ActorEntityId)
-{
-	ChangeEntityInterest(ActorEntityId, false);
-}
-
-void ASpatialFunctionalTestFlowController::ChangeEntityInterest(int64 ActorEntityId, bool bAddInterest)
-{
-	USpatialNetDriver* SpatialNetDriver = Cast<USpatialNetDriver>(GetNetDriver());
-	if (SpatialNetDriver == nullptr || ActorEntityId == SpatialConstants::INVALID_ENTITY_ID)
-	{
-		ensureMsgf(ActorEntityId != SpatialConstants::INVALID_ENTITY_ID, TEXT("Trying to change interest over an invalid entity id"));
-		return;
-	}
-
-	ULayeredLBStrategy* LayeredLBStrategy = Cast<ULayeredLBStrategy>(SpatialNetDriver->LoadBalanceStrategy);
-	USpatialFunctionalTestGridLBStrategy* TestGridLB =
-		Cast<USpatialFunctionalTestGridLBStrategy>(LayeredLBStrategy->GetLBStrategyForVisualRendering());
-	if (TestGridLB != nullptr)
-	{
-		int NumEntities = TestGridLB->Entities.Num();
-		if (bAddInterest)
-		{
-			TestGridLB->Entities.AddUnique(ActorEntityId);
-		}
-		else
-		{
-			TestGridLB->Entities.Remove(ActorEntityId);
-		}
-		if (NumEntities != TestGridLB->Entities.Num())
-		{
-			SpatialNetDriver->Sender->UpdateServerWorkerEntityInterestAndPosition();
-		}
-	}
-	else
-	{
-		UE_LOG(LogSpatialGDKFunctionalTests, Error, TEXT("Trying to change interest without USpatialFunctionalTestGridLBStrategy"));
-	}
 }
 
 void ASpatialFunctionalTestFlowController::OnReadyToRegisterWithTest()